Refact:Move methods to games, facilitate teams_hash in games
